+++ Kernel/System/Console/Command/Maint/Stats/Generate.pm
@@ -20,6 +20,7 @@ our @ObjectDependencies = (
     'Kernel::System::CSV',
     'Kernel::System::CheckItem',
     'Kernel::System::Email',
+    'Kernel::System::Main',
     'Kernel::System::PDF',
     'Kernel::System::Stats',
     'Kernel::System::Time',
@@ -369,15 +370,20 @@ sub Run {
 
     # write output
     if ( $Self->{TargetDirectory} ) {
-        if ( open my $Filehandle, '>', "$Self->{TargetDirectory}/$Attachment{Filename}" ) {    ## no critic
-            print $Filehandle $Attachment{Content};
-            close $Filehandle;
+
+        my $Success = $Kernel::OM->Get('Kernel::System::Main')->FileWrite(
+            Location   => "$Self->{TargetDirectory}/$Attachment{Filename}",
+            Content    => \$Attachment{Content},
+            Mode       => 'binmode',
+        );
+
+        if (!$Success) {
             $Self->Print("  Writing file <yellow>$Self->{TargetDirectory}/$Attachment{Filename}</yellow>.\n");
             $Self->Print("<green>Done.</green>\n");
             return $Self->ExitCodeOk();
         }
         else {
-            $Self->PrintError("Can't write $Self->{TargetDirectory}/$Attachment{Filename}: $!");
+            $Self->PrintError("Can't write $Self->{TargetDirectory}/$Attachment{Filename}!");
             return $Self->ExitCodeError();
         }
     }